refactor: organize helpers by feature
