+++ src/utils/productHelpers/index.test.ts
